getting somewhere with restaurant grouping
