+++ src/components/RestaurantCard.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Container, Divider, Header } from 'semantic-ui-react';
+import { Container, Header } from 'semantic-ui-react';
 
 const RestaurantCard = props => {
     const {id, name, city, street, zip, phone, longitude, latitude, coordinates, photo } = props
@@ -8,18 +8,7 @@ const RestaurantCard = props => {
         <>
             <Container>
             <Header size='large'>
-            <div class='restaurant_name'>{name}</div></Header>
-            {/* <p>
-            {id}
-            {item}
-            {restaurant}
-            {price}
-            {description}
-            {day}
-            {dietary}
-            {open_time}
-            {close_time}
-            </p> */}
+            <div class='restaurant_name'>Hi {name}</div></Header>
             </Container>
         </>
     )
